# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export type GatewayStatusGateway = typeof GatewayStatusGateway[keyof typeof GatewayStatusGateway];


export const GatewayStatusGateway = {
  connected: 'connected',
  unavailable: 'unavailable',
} as const;

export type GatewayStatusNetwork = typeof GatewayStatusNetwork[keyof typeof GatewayStatusNetwork];


export const GatewayStatusNetwork = {
  private: 'private',
  unknown: 'unknown',
} as const;

export type GatewayStatusSession = typeof GatewayStatusSession[keyof typeof GatewayStatusSession];


export const GatewayStatusSession = {
  active: 'active',
  required: 'required',
} as const;

export type GatewayStatusMode = typeof GatewayStatusMode[keyof typeof GatewayStatusMode];


export const GatewayStatusMode = {
  development: 'development',
  authenticated: 'authenticated',
} as const;

export type GatewayStatusPolicy = typeof GatewayStatusPolicy[keyof typeof GatewayStatusPolicy];


export const GatewayStatusPolicy = {
  enforced: 'enforced',
} as const;

export interface ResourceLimits {
  timeoutMs: number;
  maxResponseBytes: number;
  maxRedirects: number;
  maxConcurrentSessions: number;
}

export interface GatewayStatus {
  gateway: GatewayStatusGateway;
  network: GatewayStatusNetwork;
  session: GatewayStatusSession;
  mode: GatewayStatusMode;
  policy: GatewayStatusPolicy;
  limits: ResourceLimits;
}

export type GatewaySessionMode = typeof GatewaySessionMode[keyof typeof GatewaySessionMode];


export const GatewaySessionMode = {
  development: 'development',
  authenticated: 'authenticated',
} as const;

export interface GatewaySession {
  sessionId: string;
  expiresAt: string;
  mode: GatewaySessionMode;
}

export interface GatewayFetchInput {
  /** @minLength 1 */
  url: string;
}

export type GatewayFetchResponseSecurityDecision = typeof GatewayFetchResponseSecurityDecision[keyof typeof GatewayFetchResponseSecurityDecision];


export const GatewayFetchResponseSecurityDecision = {
  allowed: 'allowed',
} as const;

export interface GatewayFetchResponse {
  url: string;
  protocol: string;
  hostname: string;
  statusCode: number;
  responseTimeMs: number;
  contentType: string;
  responseSize: number;
  truncated: boolean;
  preview: string;
  redirectCount: number;
  securityDecision: GatewayFetchResponseSecurityDecision;
}

export interface PolicyBlockResponse {
  error: string;
  code: string;
  policy: string;
  message: string;
}

export type GatewayRequestSummaryStatus = typeof GatewayRequestSummaryStatus[keyof typeof GatewayRequestSummaryStatus];


export const GatewayRequestSummaryStatus = {
  allowed: 'allowed',
  blocked: 'blocked',
  failed: 'failed',
} as const;

export interface GatewayRequestSummary {
  id: string;
  timestamp: string;
  hostname: string;
  status: GatewayRequestSummaryStatus;
  /** @nullable */
  statusCode: number | null;
  durationMs: number;
  responseSize: number;
  securityDecision: string;
}

export type DashboardSummaryProtection = typeof DashboardSummaryProtection[keyof typeof DashboardSummaryProtection];


export const DashboardSummaryProtection = {
  enforced: 'enforced',
} as const;

export interface DashboardSummary {
  activeSessions: number;
  requestsToday: number;
  blockedRequests: number;
  bandwidthBytes: number;
  protection: DashboardSummaryProtection;
  recentRequests: GatewayRequestSummary[];
}