# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryProtection } from './dashboardSummaryProtection';
import type { GatewayRequestSummary } from './gatewayRequestSummary';

export interface DashboardSummary {
  activeSessions: number;
  requestsToday: number;
  blockedRequests: number;
  bandwidthBytes: number;
  protection: DashboardSummaryProtection;
  recentRequests: GatewayRequestSummary[];
}