# lib/api-zod/src/generated/types/gatewayFetchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayFetchResponseSecurityDecision } from './gatewayFetchResponseSecurityDecision';

export interface GatewayFetchResponse {
  url: string;
  protocol: string;
  hostname: string;
  statusCode: number;
  responseTimeMs: number;
  contentType: string;
  responseSize: number;
  truncated: boolean;
  preview: string;
  redirectCount: number;
  securityDecision: GatewayFetchResponseSecurityDecision;
}