# lib/api-zod/src/generated/types/gatewayRequestSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayRequestSummaryStatus } from './gatewayRequestSummaryStatus';

export interface GatewayRequestSummary {
  id: string;
  timestamp: Date;
  hostname: string;
  status: GatewayRequestSummaryStatus;
  /** @nullable */
  statusCode: number | null;
  durationMs: number;
  responseSize: number;
  securityDecision: string;
}