# lib/api-zod/src/generated/types/gatewaySession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { GatewaySessionMode } from './gatewaySessionMode';

export interface GatewaySession {
  sessionId: string;
  expiresAt: Date;
  mode: GatewaySessionMode;
}