# lib/api-zod/src/generated/types/gatewayStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayStatusGateway } from './gatewayStatusGateway';
import type { GatewayStatusMode } from './gatewayStatusMode';
import type { GatewayStatusNetwork } from './gatewayStatusNetwork';
import type { GatewayStatusPolicy } from './gatewayStatusPolicy';
import type { GatewayStatusSession } from './gatewayStatusSession';
import type { ResourceLimits } from './resourceLimits';

export interface GatewayStatus {
  gateway: GatewayStatusGateway;
  network: GatewayStatusNetwork;
  session: GatewayStatusSession;
  mode: GatewayStatusMode;
  policy: GatewayStatusPolicy;
  limits: ResourceLimits;
}