# lib/api-zod/src/generated/types/policyBlockResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface PolicyBlockResponse {
  error: string;
  code: string;
  policy: string;
  message: string;
}