# lib/api-zod/src/generated/types/resourceLimits.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Lab secure web gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceLimits {
  timeoutMs: number;
  maxResponseBytes: number;
  maxRedirects: number;
  maxConcurrentSessions: number;
}